--- a/src/adiar/bdd/apply.cpp
+++ b/src/adiar/bdd/apply.cpp
@@ -115,10 +115,7 @@ namespace adiar
 
     tpie::memory_size_type available_memory = tpie::get_memory_manager().available();
 
-    apply_priority_queue_t appD(available_memory / 2);
-    appD.hook_meta_stream(bdd_1);
-    appD.hook_meta_stream(bdd_2);
-
+    apply_priority_queue_t appD({bdd_1,bdd_2}, available_memory / 2);
     apply_data_priority_queue_t appD_data;
 
     label_t out_label = label_of(std::min(v1.uid, v2.uid));