--- a/src/adiar/bdd/quantify.cpp
+++ b/src/adiar/bdd/quantify.cpp
@@ -163,7 +163,7 @@ namespace adiar
 
     tpie::memory_size_type available_memory = tpie::get_memory_manager().available();
     quantify_priority_queue_t quantD({bdd}, available_memory / 2);
-    quantify_data_priority_queue_t quantD_data;
+    quantify_data_priority_queue_t quantD_data(calc_tpie_pq_factor(available_memory / 2));
 
     label_t out_label = label_of(v.uid);
     id_t out_id = 0;