I measured the performance on the 14-Queens problem with 4096 MiB of memory. The following is the timings I recorded.

| Before | After |
|----------|---------|
| 700     | 706   |
| 695     | 697   | 
| 699     | 697   |

Which on average is 698 against 700 seconds, which is a difference of 0.28% and completely within the error margins of our experiment.